modified movement code, imported clouds, using scroller for parallax
not including the shader in this push, as I figured using the scroller actually makes it easier to parallax multiple backgrounds with different speeds by just making a new game object for each layer. the player movement behaviour normalises input now so that diagonal inputs don't have a higher magnitude than single-axis movements.
I will bring in the custom shader+materials if it becomes necessary but for now I don't want to overcomplicate the workspace for teammates. also the ground rn is a nebula still, so i need to get a seamless water texture now